Updated copy for press and timeline pages
This bring both the press listing and the timeline pages up to date.

I did not update "News" as of yet, because I would like to recommend that this section is somewhat redundant if we have the Timeline section. There is a subtle difference between the two, but they still cover similar ground.

If maintained, the title "News" should be changed, or the word "Project" put before it. Otherwise people assume "News" means press. And the Press Link should be visible from the home page and not buried.