Update references to python-tuf repo
In preparation for the 1.0.0 release of the TUF reference
implementation python-tuf documentation linked on the TUF
website is being moved. This patch updates these links. In
particular:

- In menu item "Getting started"
Point to "Reference implementation" repo instead of "Implementation
Getting Started" page. The Getting Started page is being
deprecated, and the repo (i.e. README.md) should be just as
informative.

- In "FAQ: Are there ways to reduce bandwidth costs?"
Point to new hashed bin delegation example and specific section in
spec (old tutorial is deprecated)

- In "FAQ: 13. How can I try TUF?"
Point to API reference on RTD and up-to-date source code examples
in python-tuf (getting started is deprecated)

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>